// src/restaurant/types/restaurant/read-restaurant-data.ts
import { OrderType } from 'src/constants/orderType';

export class ReadRestaurantData {
  restaurantId: string;
  ownerId: string;
  logo: string | null;
  lbn: string;
  dba: string;
  cuisineType: string;
  storePhone: string;
  businessEmail: string;
  instagram: string | null;
  website: string | null;
  mainImgUrl: string;
  sub1ImgUrl: string;
  sub2ImgUrl: string;
  bannerImgUrl: string;
  deliveryRadius: number;
  prepTime: number;
  orderType: OrderType;
}